added in the updated set of pytests
